+++ NanoChallengeWebService/NanoChallengeWebService/Core/Movies/View/Components/GenresDesign.swift
@@ -1,0 +1,21 @@
+//
+//  GenresDesign.swift
+//  NanoChallengeWebService
+//
+//  Created by Thiago Pereira de Menezes on 26/08/23.
+//
+
+import SwiftUI
+
+//Define o Visual de todos os gêneros
+struct GenresDesign: View {
+    
+    var name: String
+    
+    var body: some View {
+        Button(action: {}) {
+            Text(name)
+        }
+//        .border(Color.blue, width: 2)
+    }
+}
